Rename varats.utils to varats.ts_utils
